requirements.txt
fastapi
uvicorn
pytubefix
git+https://github.com/openai/whisper.git
pydantic
google-genai
yt_dlp